Wine Quality Prediction (Random Forest Classifier)
A program that predicts the quality of a wine sample based on the data provided such as citric acid content, acidity levels, color, etc. The dataset used has been downloaded from Kaggle. The program uses a random forest classifier to predict the quality of the wine sample and uses seaborn heatmaps to find the correlation of the datapoints provided with the quality.